rename data class, interface and impl
