--- src/main/kotlin/org/bibletranslationtools/fetcher/repository/FileTypeCatalog.kt
@@ -1,7 +1,0 @@
-package org.bibletranslationtools.fetcher.repository
-
-import org.bibletranslationtools.fetcher.data.FileType
-
-interface FileTypeCatalog {
-    fun getAll(): List<FileType>
-}
