+++ src/main/kotlin/org/bibletranslationtools/fetcher/repository/ProductCatalog.kt
@@ -1,0 +1,7 @@
+package org.bibletranslationtools.fetcher.repository
+
+import org.bibletranslationtools.fetcher.data.Product
+
+interface ProductCatalog {
+    fun getAll(): List<Product>
+}
