+++ src/main/kotlin/org/bibletranslationtools/fetcher/repository/implementations/ProductCatalogImpl.kt
@@ -1,0 +1,49 @@
+package org.bibletranslationtools.fetcher.repository.implementations
+
+import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
+import com.fasterxml.jackson.module.kotlin.readValue
+import java.io.File
+import java.io.FileNotFoundException
+import org.bibletranslationtools.fetcher.data.Product
+import org.bibletranslationtools.fetcher.repository.ProductCatalog
+import org.slf4j.LoggerFactory
+
+class ProductCatalogImpl : ProductCatalog {
+    private val logger = LoggerFactory.getLogger(javaClass)
+    private val fileTypesResourceName = "product_catalog.json"
+
+    private data class ProductSchema(
+        val slug: String,
+        val titleKey: String,
+        val descriptionKey: String,
+        val iconLink: String
+    )
+
+    override fun getAll(): List<Product> {
+        val jsonFileTypes: String = try {
+            val productsFile = getProductCatalogFile()
+            productsFile.readText()
+        } catch (e: FileNotFoundException) {
+            logger.error("Resource File Not Found")
+            return listOf()
+        }
+
+        val products: List<ProductSchema> = jacksonObjectMapper().readValue(jsonFileTypes)
+        return products.map {
+            Product(
+                it.slug,
+                it.titleKey,
+                it.descriptionKey,
+                File(it.iconLink)
+            )
+        }
+    }
+
+    @Throws(FileNotFoundException::class)
+    private fun getProductCatalogFile(): File {
+        val resourceFileURL = javaClass.classLoader.getResource(fileTypesResourceName)
+            ?: throw FileNotFoundException()
+
+        return File(resourceFileURL.path)
+    }
+}
